Modified the constructor of the Entity class to also pass in a theta_offset. Needed to do this as stage publishes the robot's pose relative to it's initial theta position.
Also added attributes to the Entity class to keep track of the initial pose of the entity.

Fixed a bug in the Stage_Odom callback function. It was incorrectly updating the position, as it wasn't taking into consideration the initial offsets.

Finally updated the constructors of the Human, Robot and Visitor classes to match the Entity class constructor.