Fixed a bug in the StageOdom_callback method. When updating the position of the given entity, it will now take into account its initial direction as well.
Tested the changes on initial directions of North, Northwest, West and etc, and it worked correctly.

Also refactored the function significantly. The update of the px and py positions is now in a function called update_position(self, x, y), and this function is called by the StageOdom_callback method.