Added Harry's GUI stuff to the StageOdom callback function.
Also modified the random behaviour of the Visitor. The Visitor will now alternate between moving a random direction/distance, and moving to a random location.

Also modified the StageLaser callback function so that if the Visitor detects an object in front of it, it will turn right then move forward 3.